- Moved current existing UI extensions from package com.kickstarter.extensions to com.kickstarter.ui.extensions - Remove EditTextExtension.kt because there is al already existing one EditTextExt.kt added in that file the EditText.setTextAndSelection method
