+++ app/src/main/java/com/kickstarter/libs/BaseActivity.java
@@ -10,11 +10,11 @@ import com.google.firebase.crashlytics.FirebaseCrashlytics;
 import com.kickstarter.ApplicationComponent;
 import com.kickstarter.KSApplication;
 import com.kickstarter.R;
-import com.kickstarter.extensions.ActivityExtKt;
 import com.kickstarter.libs.qualifiers.RequiresActivityViewModel;
 import com.kickstarter.libs.utils.BundleUtils;
 import com.kickstarter.services.ConnectivityReceiver;
 import com.kickstarter.ui.data.ActivityResult;
+import com.kickstarter.ui.extensions.ActivityExtKt;
 import com.qualtrics.digital.Qualtrics;
 import com.trello.rxlifecycle.ActivityEvent;
 import com.trello.rxlifecycle.RxLifecycle;
