+++ app/src/main/java/com/kickstarter/ui/activities/MessageCreatorActivity.kt
@@ -3,8 +3,8 @@ package com.kickstarter.ui.activities
 import android.content.Intent
 import android.os.Bundle
 import com.kickstarter.R
-import com.kickstarter.extensions.onChange
-import com.kickstarter.extensions.showSnackbar
+import com.kickstarter.ui.extensions.onChange
+import com.kickstarter.ui.extensions.showSnackbar
 import com.kickstarter.libs.BaseActivity
 import com.kickstarter.libs.KSString
 import com.kickstarter.libs.KoalaContext
