+++ app/src/main/java/com/kickstarter/ui/extensions/ActivityExt.kt
@@ -1,4 +1,4 @@
-package com.kickstarter.extensions
+package com.kickstarter.ui.extensions
 
 import android.app.Activity
 import android.content.Context
