+++ app/src/main/java/com/kickstarter/ui/fragments/BackingAddOnsFragment.kt
@@ -9,7 +9,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.kickstarter.R
-import com.kickstarter.extensions.hideKeyboard
+import com.kickstarter.ui.extensions.hideKeyboard
 import com.kickstarter.libs.BaseFragment
 import com.kickstarter.libs.qualifiers.RequiresFragmentViewModel
 import com.kickstarter.libs.rx.transformers.Transformers
